django

+++ django_practice/02_django/articles/views.py
@@ -1,6 +1,8 @@
 from re import A
 from django.shortcuts import render, redirect
 from .models import Article
+from .forms import ArticleForm
+
 
 
 # Create your views here.
@@ -13,32 +15,26 @@ def index(request):
     return render(request, 'articles/index.html', context)
 
 
-def new(request):
-    return render(request, 'articles/new.html')
+# def new(request):
+#     form = ArticleForm()
+#     context = {
+#         'form': form,
+#     }
+#     return render(request, 'articles/new.html', context)
 
 
 def create(request):
-    # 사용자의 데이터를 받아서
-    title = request.POST.get('title')
-    content = request.POST.get('content')
-    # print('서버도착', title, content)
-
-    # DB에 저장
-    # 1
-    # article = Article()
-    # article.title = title
-    # article.content = content
-    # article.save()
-
-    # 2
-    article = Article(title=title, content=content)
-    article.save()
-
-    # 3
-    # Article.objects.create(title=title, content=content)
-    
-
-    return redirect('articles:detail', article.pk)
+    if request.method == 'POST':
+        form = ArticleForm(request.POST)
+        if form.is_valid():
+            article = form.save()
+            return redirect('articles:detail', article.pk)
+    else:
+        form = ArticleForm()
+    context = {
+        'form': form,
+    }
+    return render(request, 'articles/create.html', context)
 
 def detail(request, pk):
     article = Article.objects.get(pk=pk)
@@ -49,19 +45,31 @@ def detail(request, pk):
 
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
-    return redirect('articles:index')
+    if request.method == 'POST':
+        article.delete()
+        return redirect('articles:index')
+    return redirect('articles:detail', article.pk)
 
-def edit(request, pk):
-    article = Article.objects.get(pk=pk)
-    context = {
-        'article' : article,
-    }
-    return render(request, 'articles/edit.html', context)
+# def edit(request, pk):
+#     article = Article.objects.get(pk=pk)
+#     form = ArticleForm(instance=article)
+#     context = {
+#         'article' : article,
+#         'form' : form,
+#     }
+#     return render(request, 'articles/edit.html', context)
 
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    article.title = request.POST.get('title')
-    article.content = request.POST.get('content')
-    article.save()
-    return redirect('articles:detail', article.pk)
+    if request.method == 'POST':
+        form = ArticleForm(request.POST, instance=article)
+        if form.is_valid():
+            form.save()
+            return redirect('articles:detail', article.pk)
+    else:
+        form = ArticleForm(instance=article)
+    context = {
+        'form': form,
+        'article': article,
+    }
+    return render(request, 'articles/update.html', context)
